+++ src/templates/token/ERC20/airdrop.js
@@ -1,28 +1,28 @@
 const {constants, utils} = require('ethers');
-const {templatedMigration, buildArg} = require('../utils');
-const {tryAggregate} = require('../../../src/helpers/multicall');
+const {templatedMigration, buildArg} = require('../../utils');
+const {tryAggregate} = require('../../../helpers/multicall');
 
-module.exports = function (name, initialHolders, initialAmountsOrAmount, options = {}) {
+module.exports = function (name, holders, amountOrAmounts, options = {}) {
   const migration = templatedMigration(async (hre) => {
     const {execute, log} = hre.deployments;
 
     const [tokenName, tokenSymbol, tokenDecimals] = (
-      await tryAggregate('MultiStaticCall@0.1.0', true, [
+      await tryAggregate('MultiStaticCall@0.3.0', true, [
         {contractName: name, method: 'name', returnType: 'string'},
         {contractName: name, method: 'symbol', returnType: 'string'},
         {contractName: name, method: 'decimals', returnType: 'uint8'},
       ])
     ).map((result) => result.returnData);
 
-    const holders = await buildArg(hre, initialHolders);
-    let amounts = await buildArg(hre, initialAmountsOrAmount);
+    const recipients = await buildArg(hre, holders);
+    let amounts = await buildArg(hre, amountOrAmounts);
     if (!Array.isArray(amounts)) {
-      amounts = holders.map(() => amounts);
+      amounts = recipients.map(() => amounts);
     }
 
     log(`${name}: delivering '${tokenName}' initial airdrop:`);
-    for (let i = 0; i < holders.length; i++) {
-      log(`  ${holders[i]}: ${utils.formatUnits(amounts[i], tokenDecimals)} ${tokenSymbol}`);
+    for (let i = 0; i < recipients.length; i++) {
+      log(`  ${recipients[i]}: ${utils.formatUnits(amounts[i], tokenDecimals)} ${tokenSymbol}`);
     }
 
     const executeOptions = {
@@ -31,7 +31,7 @@ module.exports = function (name, initialHolders, initialAmountsOrAmount, options
     };
     if (options.log === undefined) executeOptions.log = true;
 
-    await execute(name, executeOptions, 'batchMint', holders, amounts);
+    await execute(name, executeOptions, 'batchMint', recipients, amounts);
   });
 
   migration.skip = async (hre) => {