+++ migrations/token/ERC20/polygon_bridging/fx-portal/FxERC20FixedSupplyRootTunnel@0.3_setFxChildTunnel_FxERC20FixedSupplyChildTunnel@0.3.js
@@ -1,0 +1,4 @@
+const setFxChildTunnel = require('../../../../../src/templates/token/ERC20/polygon_bridging/fx-portal/FxERC20RootTunnel_setFxChildTunnel');
+
+module.exports = setFxChildTunnel('FxERC20FixedSupplyRootTunnel@0.3', 'FxERC20FixedSupplyChildTunnel@0.3');
+module.exports.tags = ['ERC20', 'PolygonFxPortal'];